refactor(pedagogy): split prompt_pack.rs into responsibility submodules
The 944-line `prompt_pack.rs` is split by responsibility into a directory
module behind a flat `pub use` façade, keeping every external
`prompt_pack::<name>` path unchanged:

- `mod.rs` (206) — public trait `PromptPack`, `PackStatus`,
  `VoiceStateLabels`, and the `pub use` re-export of `load`/`load_cached`/
  `TomlPromptPack`.
- `render.rs` (146) — pure placeholder validation + brace-aware
  templating (`validate_placeholders`, `render_template`,
  `unescape_braces`).
- `intents.rs` (54) — `PedagogicalIntent` ↔ TOML-key mapping
  (`ALL_INTENTS`, `intent_key`, `intent_index`, `parse_intent_key`).
- `loader.rs` (161) — `load` / `load_cached` entry points + the
  preview-status warning gate.
- `toml_pack.rs` (400) — the `TomlPromptPack` impl, raw deserialisation
  types, and load-time validation.

The already-extracted `prompt_pack/tests.rs` gains explicit imports for
the internal helpers that now live in siblings (the `pub use` façade
cannot carry private `use` aliases or non-re-exported items).

Behaviour-preserving: identical 7-symbol external pub surface old-vs-new,
211/211 primer-pedagogy tests (unchanged from baseline), full-workspace
`cargo test` green, clippy `-D warnings` + fmt clean.
